파트너스 수정

+++ src/main/java/kr/kickon/api/admin/partners/AdminPartnersController.java
@@ -12,6 +12,7 @@ import kr.kickon.api.admin.partners.dto.PartnersDetailDTO;
 import kr.kickon.api.admin.partners.dto.PartnersListDTO;
 import kr.kickon.api.admin.partners.request.CreatePartnersRequestDTO;
 import kr.kickon.api.admin.partners.request.PartnersFilterRequest;
+import kr.kickon.api.admin.partners.request.PatchPartnersRequestDTO;
 import kr.kickon.api.admin.partners.response.GetPartnersDetailResponse;
 import kr.kickon.api.admin.partners.response.GetPartnersResponse;
 import kr.kickon.api.global.common.PagedMetaDTO;
@@ -27,6 +28,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -64,14 +66,14 @@ public class AdminPartnersController {
     );
   }
 
-  @GetMapping("/{partnersPk}")
+  @GetMapping("/{pk}")
   @Operation(summary = "파트너스 상세 조회", description = "파트너스의 상세 정보를 조회합니다.")
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "성공",
           content = @Content(schema = @Schema(implementation = GetPartnersDetailResponse.class))),
   })
-  public ResponseEntity<ResponseDTO<PartnersDetailDTO>> getPartnersDetail(@PathVariable Long partnersPk) {
-    Partners partners = adminpartnersService.findByPk(partnersPk);
+  public ResponseEntity<ResponseDTO<PartnersDetailDTO>> getPartnersDetail(@PathVariable Long pk) {
+    Partners partners = adminpartnersService.findByPk(pk);
     if (partners == null) throw new NotFoundException(ResponseCode.NOT_FOUND_USER);
     PartnersDetailDTO dto = adminpartnersService.getPartnersDetail(partners);
 
@@ -89,14 +91,29 @@ public class AdminPartnersController {
     return ResponseEntity.ok(ResponseDTO.success(ResponseCode.SUCCESS, responseDto));
   }
 
-  @DeleteMapping("/{partnersPk}")
+  @DeleteMapping("/{pk}")
   @Operation(summary = "파트너스 삭제", description = "파트너스를 삭제합니다.")
   @ApiResponses({
       @ApiResponse(responseCode = "200", description = "성공"),
   })
-  public ResponseEntity<ResponseDTO> deletePartners(@PathVariable Long partnersPk) {
-    Partners partners = adminpartnersService.findByPk(partnersPk);
+  public ResponseEntity<ResponseDTO> deletePartners(@PathVariable Long pk) {
+    Partners partners = adminpartnersService.findByPk(pk);
+    if (partners == null) throw new NotFoundException(ResponseCode.NOT_FOUND_USER);
     adminpartnersService.deletePartners(partners);
     return ResponseEntity.ok(ResponseDTO.success(ResponseCode.SUCCESS));
+  }
+
+  @PatchMapping("/{pk}")
+  @Operation(summary = "파트너스 수정", description = "파트너스를 수정합니다.")
+  @ApiResponses({
+      @ApiResponse(responseCode = "200", description = "성공",
+          content = @Content(schema = @Schema(implementation = GetPartnersDetailResponse.class))),
+  })
+  public ResponseEntity<ResponseDTO<PartnersDetailDTO>> patchPartners(@PathVariable Long pk,
+      @RequestBody PatchPartnersRequestDTO request) {
+    Partners partners = adminpartnersService.findByPk(pk);
+    if (partners == null) throw new NotFoundException(ResponseCode.NOT_FOUND_USER);
+    PartnersDetailDTO responseDto = adminpartnersService.patchPartners(partners, request);
+    return ResponseEntity.ok(ResponseDTO.success(ResponseCode.SUCCESS, responseDto));
   }
 }
